Functie genereer_zorgdata verhuisd naar GenCareAIFunctions.py
Aanpassing u_role met de onrustscore
